feat: added CLI and doc to API client

+++ api-client/pangea_api/analysis_result.py
@@ -152,6 +152,7 @@ class AnalysisResultField(RemoteObject):
         self.name = field_name
         self.stored_data = data
         self._cached_filename = None  # Used if the field points to S3, FTP, etc
+        self._temp_filename = False
 
     def nested_url(self):
         return self.parent.nested_url() + f'/fields/{self.name}'
@@ -183,7 +184,7 @@ class AnalysisResultField(RemoteObject):
         blob = self.knex.post(f'{self.canon_url()}?format=json', json=data)
         self.load_blob(blob)
 
-    def download_file(self, cache=True):
+    def download_file(self, filename=None, cache=True):
         """Return a local filepath to the file this result points to."""
         if self.stored_data.get('__type__', '').lower() != 's3':
             raise TypeError('Cannot fetch a file for a BLOB type result field.')
@@ -195,15 +196,18 @@ class AnalysisResultField(RemoteObject):
             url = self.stored_data['uri']
         if url.startswith('s3://'):
             url = self.stored_data['endpoint_url'] + '/' + url[5:]
-        myfile = NamedTemporaryFile(delete=False)
-        myfile.close()
-        urlretrieve(url, myfile.name)
+        if not filename:
+            self._temp_filename = True
+            myfile = NamedTemporaryFile(delete=False)
+            myfile.close()
+            filename = myfile.name
+        urlretrieve(url, filename)
         if cache:
-            self._cached_filename = myfile.name
-        return self._cached_filename
+            self._cached_filename = filename
+        return filename
 
     def __del__(self):
-        if self._cached_filename:
+        if self._temp_filename and self._cached_filename:
             os.remove(self._cached_filename)
